read simple persist-type reader out of vault
