+++ pkg/pillar/containerd/containerd.go
@@ -30,7 +30,7 @@ import (
 	"github.com/lf-edge/edge-containers/pkg/resolver"
 	"github.com/lf-edge/eve/pkg/pillar/base"
 	"github.com/lf-edge/eve/pkg/pillar/types"
-	"github.com/lf-edge/eve/pkg/pillar/vault"
+	"github.com/lf-edge/eve/pkg/pillar/utils/persist"
 	"github.com/opencontainers/go-digest"
 	"github.com/opencontainers/image-spec/identity"
 	"github.com/vishvananda/netlink"
@@ -97,7 +97,7 @@ func GetServicesNamespace() string {
 func init() {
 	logrus.Info("Containerd Init")
 	// see if we need to use zfs snapshotter based on what flavor of storage persist partition is
-	if vault.ReadPersistType() == types.PersistZFS {
+	if persist.ReadPersistType() == types.PersistZFS {
 		defaultSnapshotter = types.ZFSSnapshotter
 	}
